# Improve Guid parsing performance for "D", "N", "B", and "P"