@stephentoub 2021-07-16
@uweigand, I would not be at all surprised if I got the endianness wrong here.  Can you take a look?