@DrewScoggins 2021-07-27
We are seeing this improvement in the lab as well.

https://github.com/dotnet/perf-autofiling-issues/issues/27