@DrewScoggins 2021-07-29
Also improved on Arm64. https://github.com/dotnet/perf-autofiling-issues/issues/75